requirements.txt
selenium>=4.15.0
yadisk>=3.1.0
python-dotenv>=1.0.0
streamlit>=1.29.0
requests>=2.31.0
webdriver-manager>=4.0.1
pypdf>=5.2.0
